requirements.txt
flask===3.0.3
numpy==1.22.3
Pillow==9.4.0
matplotlib==3.6.3
pandas==1.5.2
tensorflow==2.12.0
keras==2.12.0